requirements.txt
apispec==3.3.2
attrs==20.2.0
Babel==2.8.0
click==7.1.2
colorama==0.4.4
defusedxml==0.6.0
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-AppBuilder==3.1.0
Flask-Babel==1.0.0
Flask-BabelPkg==0.9.6
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-MySQLdb==0.2.0
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.24.0
mysqlclient==2.0.2
numpy==1.19.3
pandas==1.1.4
Pillow==8.0.1
prison==0.1.3
PyJWT==1.7.1
PyMySQL==0.10.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
typing-extensions==3.7.4.3
urltools==0.4.0
Werkzeug==0.16.1
WTForms==2.3.3
XlsxWriter==1.3.7
zipp==3.4.0
